+++ packages/datasource-toolkit/src/interfaces/schema.ts
@@ -24,6 +24,8 @@ export type CollectionSchema = {
 export type RelationSchema = ManyToOneSchema | OneToManySchema | OneToOneSchema | ManyToManySchema;
 export type FieldSchema = ColumnSchema | RelationSchema;
 
+export type ColumnSchemaValidation = Array<{ operator: Operator; value?: unknown }>;
+
 export type ColumnSchema = {
   columnType: ColumnType;
   filterOperators?: Set<Operator>;
@@ -33,7 +35,7 @@ export type ColumnSchema = {
   isReadOnly?: boolean;
   isSortable?: boolean;
   type: 'Column';
-  validation?: Array<{ operator: Operator; value?: unknown }>;
+  validation?: ColumnSchemaValidation;
 };
 
 export type ManyToOneSchema = {